Add a MakeDefault trait.
The MakeDefault trait allows construction of a type T if it either
- is default-constructible
- or has a static with_default() method that returns type T.

MakeDefault<T>::has_trait reports if T has the trait. A type
must be either default-constructible or provide with_default()
but can not be both in order to implement the trait.

MakeDefault<T>::make_default() returns a T constructed with its
default value, and requires that T is MakeDefault.