Add sus::option::Option type.
The type has clear inspiration from Rust:
https://doc.rust-lang.org/std/option/enum.Option.html

This introduces the minimum methods to make the type useful and we'll
add more in further commits.

There's a question about whether we should push toward objects being
consumed as rvalues more like Rust does (with &&-qualified methods)
or not, which this CL does not answer. For now, it's matching Rust more
than normative-till-now C++.