Add Option map() map_or() and map_or_else() functions
These functions map the value in the Option through a function,
returning a new Option holding the type returned by that function.

The map() function maps None to None.
The map_or() function takes a default value to use when the Option
is None, so it always returns Some.
The map_or_else() function takes a default-value-generator function
to use when the Option is None, so it always returns Some.