Cleanup Option and make it moveable
Always use ::sus to access the sus namespace.

Don't use constexpr on functions where one path is never
constexpr (except for panic).

Use if/else instead of ? for better codegen. For example
you have to move() stuff inside a ternary operator, they
don't get treated as xvalues in the same way as when
they are the only thing in a return statement.