Add Option insert() and get_or_insert{_default,_with}()
insert() will replace whatever value is there, if any.

get_or_insert() will get a mutable reference to the value,
and insert a value if there is none in order to be able to
create a reference. This is the lvalue equivalent of
unwrap_or().

get_or_insert_default() is similar, but you don't specify
the value at the callsite. This is the lvalue equivalent of
unwrap_or_default().

get_or_insert_with() is similar too, but you give a function
to construct the value to insert, which avoids expensive
constructions. This is the lvalue equivalent of
unwrap_or_else().