Add Option as_ref() and as_mut().
These return an Option<const T&> or Option<T&> respectively, and
are called on lvalues only. This allows you to get a reference
and unwrap it, or work with it in an Option, and avoids the need
for null pointers.

Option stores _either_ a T or a T&, which avoids a union of the
both, so that Option<T&> can be pointer-sized (plus the bool)
and Option<T> can be T-sized (plus the bool).