Add Option shortcuts for as_ref().unwrap() and friends
as_ref().unwrap() -> unwrap_ref().
as_ref().expect() -> expect_ref().
as_mut().unwrap() -> unwrap_mut().
as_mut().expect() -> expect_mut().

We do this because C++ moves can be expensive and this avoids
a bunch of move codegen, so it might be useful. But we still
have the flexability to get a reference that might be None.

as_ref()/as_mut() work on lvalues, however:
- as_ref() works only on const lvalues (and not rvalues)
- as_mut() works only on mutable lvalues

The same is true for unwrap_ref/mut and expect_ref/mut.