Make use of requires to make Option<T> more-trivial when T is
If T is trivially copyable, moveable, or destructible, then we
can make Option<T> be the same. This also means that if T is
trivially relocatable, Option<T> is as well.

However, if T is trivially relocatable through the use of
[[clang::trivial_abi]], the Option<T> will not gain the same
attribute. We leave this for future work. Since attributes are
not conditional, we would likely need to split up the Option
class into OptionBase and an empty Option subclass, with two
specializations, one with the attribute and one without.