Make sizeof(Option<T&>) == sizeof(T*).
This is done by template specialization of the storage in the
option. For a reference type, Option holds a pointer and uses
the pointer value to determine if it is holding Some or None.

We double up the tests to cover the reference-type branches in
all of the Option methods.