Add Option iter(), iter_mut(), into_iter().
We introduce an Iterator trait, which has a single abstract function,
next() that returns Option<Item>. We use virtual because iterators
will be able to build on other iterators, and chain together. But
each iterator should just hold an `Iterator<Item>` within it to
avoid blowing up the numbers templates, which would happen if the
Iterator's type referred to its subclass (ie. if we used CRTP
instead).

The Option's Iterator impl initializes the first Item, and then
always returns None from next().

The Iterator type has begin() which returns an IteratorStep that
holds the current Item. It starts with the first item. And it has
end() which is a sentinel. When next() returns None, then the
operator==(end) will be true.