Reorganize iterator code and implment Iterator::all()
Avoid cyclical deps between Option and Iterator. This is tricky
because Iterator returns an Option, and Option returns an Iterator
over its contents.. which has an Option inside it.

We do this through the magic of templates, since functions are not
instantiated until they are called. So Option forward-declares its
iterator, and the user must #include "sus/traits/iter/iterator.h"
(maybe that's too deep a path) to use the iterator. This is also
required anyway to use provided methods on iterator, as we give
only a definition of provided methods in transitive includes of
the Iterator trait.

Split Iterator into a definition and function headers. Transitive
includes from other library headers will only pull in the definition
to keep include sizes down. Users of iterators will need to include
the iterator.h header (which brings in both) to use provided methods.
This can be considered a case of IWYU, though it's for a base class
here.