Add a test template for testing type properties, use it for option
We use the test template to find differences between T and
Option<T>, and fix those differences so that Option<T> can
be used in the same ways (and only the same ways) as T can.
For example, if T is moveable but not copyable, then Option<T>
should be also.

This uncovered an issue where a non-moveable type was being
move-constructed into Option's storage. So add requires
on Option's storage and a const& override such that non-
move-constructible types, even if given as an rvalue, get
copied instead.