Add Iterator filter() method.
This required a big rearchitecture of how Iterator is implemented:

A given iterator class will still inherit from something to get the
provided methods, and allow them to make use of its required next()
method. They can override methods like count() if they have a faster
implementation of them. That class is now called IteratorBase.

Now the iterator class is wrapped in a subclass of it, called
Iterator<Subclass>. And "adaptor methods" are provided there, that is
iterator methods that return another iterator, nesting the current
iterator inside it.

In Rust, this can all be done in the Iterator trait (the IteratorBase
class effectively), because type resolution is done in a more lazy
manner. So for instance:

```
    fn filter<P>(self, predicate: P) -> Filter<Self, P>
    where
        Self: Sized,
        P: FnMut(&Self::Item) -> bool,
    {
        Filter::new(self, predicate)
    }
```

This method is part of Iterator, but takes `self` by value, making it a
sized type, even though its size is determined by what thing is
implementing the trait (the subclass in C++ terms).

However in C++, the base class's type definition can't use the
subclass's definition (such as calling sizeof()) on it, only its type
name. We move the adaptor methods into a subclass Iterator<T> to be
able to see the full definition of T.

So for example, the Once<int> iterator becomes:
  Iterator<Once<int>> -> Once<int> -> IteratorBase<int>

This allows us to avoid malloc for every time we chain iterators
together, as long as each iterator can be trivially relocated. If so,
we move construct the iterator into an untyped buffer in the next
link in the iterator chain, which allows us to erase the type and limit
the impact on the compiler.

The result is that you can do something like:

```
  auto array = ....;
  auto it = array.iter()
                  .filter([](const int& i) { return i >= 3; })
                  .filter([](const int& i) { return i <= 4; });
  for (int i: it) {
    print(i);
  }
```

To allow templated classes to be trivially relocated, we need a more
expressive way to tell that a class is trivially relocatable. Clang
has [[clang::trivial_abi]] but it's an attribute on the class, so it
applies to all template instantiations. GCC and MSVC have nothing of
the sort.

We introduce a few macros that can be inserted into (the private section
of) a class to declare that it is or is not trivially relocatable, which
allows it to be memcpy'd to a new location, skipping both the move
constructor + destructor along the way.

The macros are currently in __private because they are for internal use
but I think they should move out to the public `sus::mem` namespace.
Because marking a class incorrectly, when the move constructor does need
to run on a move, would lead to UB, they all take a
`sus::marker::UnsafeFnMarker` as their first argument. Marking a class
with these will make them trivially relocatable in the same way as
[[clang:trivial_abi]] does, for library usage. It does _not_ tell the
compiler about it, and thus does not impact passing the type as an
argument. It would be nice to see the trivial_abi concept expanded to
better support this in the compiler.

sus_class_trivial_relocatable(unsafe_fn)
- Marks the class as trivially relocatable.

sus_class_trivial_relocatable_value(unsafe_fn, bool)
 - Marks the class as trivially relocatable if the bool argument is
   true.

sus_class_trivial_relocatable_value(unsafe_fn, bool)
 - Marks the class as trivially relocatable if the bool argument is
   true.

sus_class_maybe_trivial_relocatable_types(unsafe_fn, types...)
 - Marks the class as trivially relocatable if all the types passed as
   arguments are (marked as) trivially relocatable.

sus_class_assert_trivial_relocatable_types(unsafe_fn, types...)
 - Marks the class as trivially relocatable, and statically asserts that
   all the types passed as arguments are (marked as) trivially
   relocatable.