Support hiding the Option<T> flag inside the T object.
The Fn, FnMut, and FnOnce types indicate that their enum
state is never 0, and the Option<Fn> is thus the same size
as a Fn.

We add a NonNull<T> which is a never-null T*, and which indicates
that its pointer field is never zero as a result.

Then we implement Option<T&> on top of NonNull<T>.