Generate documentation from the code with hdoc (#80)
The documentation is going to https://docs.hdoc.io/danakj/subspace/ at this time.

In the future, we could split the documentation generation into a separate CI action entirely.
Since it only requires CMake configure, not the full compilation, it would not be too costly to
do so. Then we can track doc-building success separately from compile/tests success.