Avoid UB in Option unions, but we have to drop constexpr from a lot of Option methods
To avoid UB we add a standard layout type into the union when doing the NeverValue
field optimization, and read the NeverValue field from there. The type has the
same prefix as the actual type, so this is legal.

This means we restrict the NeverValue field optimization to types that are standard
layout. And since we can not read the inactive field in a union from a constant
expression, clang warns for Option methods that read the state in the NeverValue
field if the Option is set to Some. So we remove constexpr from the API to avoid
inconsistent behaviour between Option<A> and Option<B> or Option<A> and Option<A&>.