Support non-trivially-relocatable types for iterator filtering
Have the SizedIterator move the iterator being filtered onto the heap
if it's not trivially relocatable. Then the Filter iterator doesn't
need to use RelocatableStorage, since that was its job as well.

We can't just use RelocatableStorage to do this because the
SizedIterator needs to be moveable to be returnable, but since
it doesn't know its type it has to be trivially relocatable.
At that point, RelocatableStorage doesn't buy anything for the
SizedIterator type.