Add Iterator collect() method
The collect method is templated and must be given the type of
something that satisfies the sus::iter::FromIterator concept.

The FromIterator concept is satisfied by a static constructor
method called from_iter() that receives an IteratorBase rvalue.

We implement FromIterator on Result, so that an
Iterator over Result<T, E> can be collected into a
Result<C<T>, E> where C itself satisfies FromIterator. The
result can be collected into like:
  sus::move(iter_over_result_t).collect<Result<Vec<T>, E>()
If there was a Vec type that could be constructed from an
Iterator over T.