Add the Vec type
Vec has the same methods as Slice, but with additional
methods too. We added reserve() and friends, clear() and
push() so far.

The reserve() method acts like Rust's Vec and not like
std::vector. To give something that could be used in
pace of std::vector::reserve, however, we provide a
Vec::grow_to_exact() which takes a total capacity as
its input (like std::vector) instead of an additional
number of elements to ensure there's space for (like
Rust Vec::reserve()).