Add data_size_of<T>(), the size of a type T's fields excluding padding
This allows one to memcpy a trivially relocatable type into
place without potentially clobbering other objects located
in the type's padding.

This was based on the insights of @ssbr, which are desribed in the
mem/size_of.h header. However he wished that subclasses and
[[no_unique_address]] would always reuse padding in the same ways but
this is not the case with empty classes and the MSVC compiler. This
is also described in mem/size_of.h.