Make the panic functions overridable at compile time.
This allows the user of subspace to provide their own implementation
of these methods.

Note that we choose to do this at compile time instead of linking
time so that the resulting choice can be smaller than a function call:
https://discourse.llvm.org/t/rfc-c-buffer-hardening/65734/20

Additionally the libc++ approach of using __attribute__((weak)) does
not work on Windows, but a compile-time approach does.