Add support for structured bindings to Tuple
This requires not having a get() method, because we don't
overload get() for const&, & and &&. But the presence of get()
causes the compiler to try call it. So we rename get() to
get_ref().

Then we add get() in the same namespace as a free function with
all three overloads for structured bindings to fall back to and
use. Much like begin() and end() for ranged-for-loop support
from Iterator