Fix structured bindings with Tuple and move
The bindings will unwrap<I>() each element in the tuple, which
was being treated as a use after move. However since each
element in the tuple can be moved from independently, this is
really okay... just complicated. So replace the bool moved-from
with a uint64_t bitmap that tracks the moved-from state of the
first 64 elements.