Prevent use-after-move of a Tuple when the whole tuple is moved
To help support this, move the use-after-move methods into a
marker type so the logic is centralized. Then do the use-after-
move checks for the whole tuple being moved in the move and
copy operations for that marker type so we don't have to define
any copy/move operations for the Tuple.