Implement Ord methods and simplify Eq and Ord concepts
Give a default value to the second type variable for the Eq and Ord
concepts so that Eq<int> is true and you don't have to write
Eq<int, int>. However we still do need the second type in the concept in
order to allow comparison between things like int and const int or int
and int& due to a lack of blanket trait impls (like
https://doc.rust-lang.org/stable/std/cmp/trait.Eq.html#impl-Eq-for-%26A)
and the fact that references and qualifiers produce different types in
C++.

Template types like Option or Tuple use a template argument to allow
compare between Option<T> and Option<const T&>, etc.

Implement and test min/max min_by/max_by and min_by_key/max_by_key
functions in the sus::ops namespace. They require the type that ends up
being compared to satisfy Ord. We receive and return `T` instead of
`const T&` to avoid introducing references and lifetimes when the input
was not a reference. This allows moving 2 objects through min(a, b) and
keeping only one of them.

It would be nice to receive `T&&` and return `T` thus ensuring
references to temporaries are never returned, but inputs could be `const
T&` and then would return `const T&`. However then if one input is const
and one is not, the type `T` becomes non-deducible. We could provide
overloads, where one branch clones the input. This puts an extra
requirement (i.e. Clone) on the types in the case one input is a const
and the other isn't, which is confusing. Furthermore, we can't use a
forwarding reference, so we need overloads for all combination of const
T&, T&, and T&&, for each parameter, which is a huge combinatorial
explosion of overloads. Instead, we rely on the compiler to optimize out
the move.

Users can explicitly request to receive and return a reference by
specifying the type parameter, like `sus::ops::min<i32&>(a, b);`.