Allow explicit conversions from integers to primitives
This is important because operators on primitive values of size < int will
be promoted to int. This makes it very hard to do something like
size_t{~my_u16} without this conversions, because if you write
size_t{~my_u16.primitive_value}, you're negating the uint16_t which
will promote to int, which is now signed and not what you wanted at
all.