Mark the panic functions with attributes to keep crashes distinct
The not_tail_called ensures that there's a distinct call to panic()
rather than being inlined into the bottom of another function.

The nomerge ensures that a call to panic() doesn't get merged with
a different call to it in the same context.

h/t to @pbos for review and inspiring me to use these attributes.