Add Option::unzip()
In the process, we found we needed much better support for holding
references in Union and Tuple, which is now done.

into_inner() will return the correct reference type, and we test for
this. And forward_mref() will call mref() on an lalue reference but not
touch other types, for forwarding along a mutable reference return type.