Require Union tags to be Copy.
The which() function operates by constructing a constexpr array
of tags, and the returning from there. The clone() needed to
construct that array would be done at compile time (but may cause
slow builds). However as the array is const(expr) it would have
to be cloned out of the array to be returned, and we don't want
switch (my_union) {} to have a secret performance cliff. So we
indicate the tag will be copied in the API by requiring the type
to be Copy.