Rationalize references with Copy/Move/Clone
The Copy, Clone and Move traits are no longer true for all references,
instead they report back on the object type behind the reference. This is
normally the right thing, as copying from a reference will copy from the
underlying object, etc.

The clone() function returns an object type even if passed a reference.

We add new CopyOrRef, CloneOrRef, and MoveOrRef concepts which are true
for all references. These are used in templates that are generic over
references, and require the user to take care to act on the reference
and not hte underlying object. We use these to implement Option.

The move() function must accept any input, because even non-movable
types can have rvalue-qualified methods.

We add clone_or_forward<T>() that will return a T, which can be a reference
type, this is like sus::forward<T> but it clones if T is not a reference
instead of casting to an rvalue.

In the process I noticed Tuple<T&> was not Copy, which is incorrect. Fix
this by specializing the Storage types for references, storing a pointer
instead of a reference internally, which is then Copy.