Add a SameOrSubclassOf concept
The concept helps avoid array decay to a pointer by comparing pointer
types inside the concept. This encourages a function to receive a type
T and not T*, as the latter decays from an array. Then the concept is
false for an array type as it requires pointer types.

Use SameOrSubclassOf to replace std::is_convertible_v for NonNull<T>.