Add CMake option for building CIR.
This CL adds an option into the cmake file to enable/disable the
building of CIR. The default is `ON` but it will also honour the
`SUBSPACE_NO_BUILD_CIR` environment variable to default off if set.

This allows doing `cmake -GNinja -DSUBPSACE_BUILD_CIR=OFF ../..` in
order to configure cmake with the flag disabled (it would also allow
setting the option in a downstream CMake configuration file).