Add Option operator* and operator-> for lvalues
Working with lvalues is really difficult in C++ if you have to always
explicitly move() them to use them. The unwrap_ref() and unwrap_mut()
methods were really quite loud too. as_ref().unwrap() is no better _and_
can't be constexpr due to as_ref() needing to query the state of the
Option.

I was starting down some dark places with lambdas and it's not worth it.
Loosen the rules to allow overloading for operator* and operator-> since
they can't be spelled without it.

This seems like a spot where new users would very quickly bounce off
subspace and it's extremely idoimatic for C++, so we should provide
them.

Consider them for result too.