Make all of Option constexpr
To do this we have to make the implementation of NeverValueField for
a given type a bit more awkward.

1. The type needs to be trivially default constructible. It can be
   private, since this implies leaving its fields all uninitialized. The
   NeverValueAccess class will still be able to construct it.
2. The macro takes an additional 'destroy value'.
3. The destructor must be a no-op when the never value field is set to
   the destroy value. Most times the never value and destroy value can
   be the same thing but if the destructor branches on the value then it
   can be set to a no-op value.

Once these were done then we introduced the NeverValueAccess as a type
that holds a NeverValueField type T, and when default constructed it
trivially constructs the T. The caller must then set the never value in
the newly constructed T. Default construction is required to be able to
destroy the member in Option and reset it to the never value in a
constexpr way (without in-place new).

The good news is there's no longer any standard-layout requirements.

The bad news is that this probably dooms the NeverValueField concept to
be unstable, or worse, private. Implementing it in user types is just
kind of messy and complicated.

The other bad news is Vec no longer satisfies NeverValueField, since
it has a public, non-trivial default constructor. So it's not possible
to construct it with all its fields uninitialized.