Improve compile errors with Fn of function pointer
When storing a function pointer the types must match
exactly. We were allowing construction of Fn with a
pointer of different types, and then we'd fail to
compile deep inside the Fn implementation.

Implement the FnPointerMatches concept to ensure the
types in the function pointer match exactly and use
that as the bound in the Fn APIs.