Improve compile errors with Iterator::map()
If passing in a type that can't convert to FnMut<R(Item&&)>,
then avoid participating in overload resolution instead of
failing to compile when later trying to construct the FnMut.

We do this by adding a bound on Into<MapFn, FnMut<R(Item&&)>,
and we sus::into() the input to construct the FnMut.