Normalize the Choice API
Add as<Tag>() and as_mut<Tag>() for checked access
Add get<Tag>() and get_mut<Tag>() for Optional access
Add get_unchecked<Tag>() and get_unchecked_mut<Tag> for unsafe access.

Allow access to reference types when Choice is an rvalue but not to value types.

Introduce a .construct() on Tuple' marker type when you want
to not write out the whole type signature but you need the
concrete type (in this case because it's going to be received
as a templated const reference and then it can't be converted
later.