Have FromIterator receive an IntoIterator type.
Instead of receiving an IteratorBase<Item> type. This allows us to
eliminate IteratorBase and drop virtual on Iterator::next().

The downside is the template implementation of Option::from_iter()
and Result::from_iter() is much more complex, as you can't deduce
the type inside a concept you're receiving as a parameter.