Allow FnOnce to be split into any number of FnOnce
The splitting mechanism just holds a reference to the original FnOnce
so two calls to different splits of the FnOnce will panic.

The splittinf mechanism also tracks the calls in debug mode to provide
a more clear crash indicating two calls were made, instead of just that
the FnOnce was moved from (which is what happens when it's run).