Rename Fn to FnRef, FnMut to FnMutRef and FnOnce to FnOnceRef
These are concrete impls of what will be the Fn/FnMut/FnOnce concepts
that are a light reference around the thing they call while erasing its
type to avoid requiring a template to receive and use a callable thing
in a function (that doesn't store the callable).

They are however not constexpr due to language limitations so in
constexpr vocab types, the concepts will need to be used directly.