Fn, FnMut, FnOnce concepts take a signature
Instead of Fn<Return, Arg1, Arg2> you now write Fn<Return(Arg1, Arg2)>

And when it is passed something that can't be called with (Arg1, Arg2)
the failing includes the type 'NoOverloadMatchesArguments' to indicate
the failure.

If the return type doesn't convert to Return, the usual and best concept
error message is reported where the resolve return type is compared
against the desired one.