Define more clearly when to make something Copy
The Slice type is 2 pointer-sized and would typically be Copy in
C++. In Rust, a slice is not Copy, but it is passed by reference which
is (when not mutable).

We want to avoid unnecessary pain when using core types like Slice.
Moving is a papercut, but is needed for performance reasons, but when
not justified by such, we should not require it. Making Slice be Copy
makes sense since:
a) moving and copying copy the same amount of data.
b) it's fine and encouraged to pass Slice by value, as its size is <=
   2 pointers.

This also allows passing slices to have the same semanics as a Rust
(non-mut) slice, as it can be passed to a call without a move taking
place, and used afterward.