Add Slice::chunks() and Slice::chunks_mut()
To implement these, we also add Slice::split_at_unchecked() and
Slice::split_at_mut_unchecked().

The chunks methods return an iterator over subslices that will each
be the requested size, except for the last chunk which may be smaller
if the size doesn't divide into the Slice's total size.