Split Slice into Slice/SliceMut, give Vec all of Slice/SliceMut methods
We use the slice macros to give all the Slice methods to SliceMut, and
all the methods of both slice types to Vec. However Vec deletes rvalue
usage of methods that return references to its internals.

Vec is convertible to a SliceMut& or Slice&, which requires Vec to hold
a SliceMut internally. Same for SliceMut converting to a Slice&. Then
they must be friends in order to reach into slice and set fields.