Avoid ambiguous conversions to Slice from rvalues
Converting from && to Slice&& is ambiguous with returning a const Slice&
as the Slice type has to pick a copy or a move constructor.

Instead we can convert from && to Slice (not a reference) which avoids
needing to pick a constructor on the Slice type, we chose it in the
conversion operator (the move one).