Add Slice::copy_from_slice, copy_from_slice_unchecked, clone_from_slice
The copy_from_slice() comes with a number of checks: that the slices
don't overlap, they have the same length, and the number of bytes does
not overflow. So we add a copy_from_slice_unchecked() as an unsafe
performance backdoor. These methods only work for types that are
trivially copyable.

clone_from_slice() is used for types that are not trivially copyable.