Add ptr::copy and copy_nonoverlapping
Use these instead of memcpy. They do the conversion to bytes for
you, and require you to explicitly cast to chars if that's what you
want to do. They also panic instead of causing UB if the number of
bytes overflows. And provide debug checks to try catch bugs in debug
builds.