Place stream support directly near the fmt::formatter
This ensures that it is not possible for gtest users to end up
including the type and printing it in one TU without stream support
which then breaks printing it in another TU that also includes the
stream support. They will simply always be printable by streams.
This is done without including anything other than iosfwd (and string
and fmt) in header files.