Use macro expansion location for file include/exclude
When visiting the use of a macro, find the place where the
macro is expanded into and use that as the location for deciding
if we're in a file we want to include things from.

The macro location itself has no file at all, as its a temp
buffer generated by the compiler, whereas the expansion site
will be a real file and always has a FileEntry.

This fixes a crash that was occuring in third_party/fmt/ headers
when excluding `/third_party/` paths.