Fix format_to_stream for GCC
GCC will resolve concepts on overloads where other parameters
already don't match the caller. This results in recursive concept
resolution. It only does this if the other parameters are _not_
templates tho. So in this case we need to shortcut before the
concept resolution with std::same_as, to discard the overload
before GCC gets a chance to look at the concept.

Apparently this is all unspec'd behaviour and up to the compiler
to do what it likes, and GCC currently prefers this.

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=99599