Rename Option::some() to Option::with()
We get into a weird naming thing otherwise with Result. We'd
like to construct Result as Result::ok(x) if we can construct
Option as Option::some(x). However Result has a method `.ok()`
which converts it to an Option. This points to the namespace
of useful methods being taken by ctor methods being a problem
and thus why we want to use `with()` or `with_x()` as our ctor
method names. This brings Option into consistency with all the
other types.