Move iterator ctors to private for types constructed by Iterator methods
This is an attempt to help disambiguate the types that are meant to be
constructed by users of the library and those which are not.

It would be nice to have free functions to construct the types users
use but unfortunately type deduction of template types sucks so it does
not work as well as constructor methods, which have the type being
iterated over spelled out in a more consistent way - in the iterator's
class type - instead of as an explicit function type parameter.