Add Iterator::by_ref()
It is callable on an iterator without consuming it, and it
returns another iterator that refers to the first. The
new iterator can be chained (e.g. rev() or filter()) and
it will pull elements from the same place as the original
iterator, but the chaining will not affect the original
iterator. They both become empty at the same time.