Plumb size_hint() through iterator chains
To do this, size_hint() becomes a required method on the Iterator
concept. However that means an Iteration defined in an Option
needs to know about SizeHint which... uses Option.

Fix circular includes by moving all Option and Tuple using
methods on unsigned integers into the impl header, and moving
the includes of option.h and tuple.h there. This allows the
Option headers and SizeHint header (included from Option's
header now) to include unsigned integers.

Then also move the impl of the Option method that defines an
iterator to the bottom of the file, and include size_hint.h
there only after all of Option is defined, since it will try
to include option.h back and can't.