Rework NeverValueField to avoid double destruction
If a type had non-trivial move assignment, then doing
```
t.~T();
t = T();
```
would not just reset the union to an invalid but active
state, but would effecively destroy `t` twice as the assignment
may see valid state left behind by the destructor and try to destroy
it again.

Replace the trivial default constructor requirement with an explicit
constructor from a NeverValueConstructor marker type, which inits the
object into the never-value state, and remove the set_never_value()
code paths. This makes objects have to be a bit more uh.. helpful
with never-values outside of the macro, but it's needed for constexpr
anyways.

Add a constexpr path to do the above reset without depending on a
trivial move assignment. But we also don't want to depend on the
in-place new operator, as that isn't constexpr. So types can opt
into a constexpr path by providing `destroy_and_set_never_value()`
as a private function. This should act as if the destructor was
called and then the constructor from the NeverValueConstructor marker
type.

This fixes a double-free crash when chaining VecIntoIterator which
gets erased into SizedIterator and which has a move assignment that
is non-trivial.