All Iterator compare (cmp/eq/etc) methods to compare with other types
Instead of requiring the other iterator to have the same types
exactly, allow comparison with an iterator as long as the types
they vend are compatible, ie. Eq, Ord, PartialOrd, or WeakOrd.

Iterator find methods don't need the iterator to be trivially
relocatable. They consume the iterator and don't need to store it into
another iterator.