Catch iterator invalidation for Vec and Array
The invalidation checks notice when a SliceIter or SliceMutIter
is active when the Vec is mutated or the Vec/Array is moved.

We use [[msvc::no_unique_address]] for disabled Iterator references so
they disappear from object sizes if invalidation tracking is off. This
uncovered an MSVC bug however where it gets the alignment wrong for the
outer type if the [[msvc::no_unique_address]] field comes before an
array. Filed upstream, but for that reason we can use it in Slice/Vec
and in iterators, but not in Array.
https://developercommunity.visualstudio.com/t/MSVC-produces-incorrect-alignment-with-/10416202

clang-cl doesn't support msvc::no_unique_address, so expect additional
padding there as well: https://github.com/llvm/llvm-project/issues/49358

We no longer allow Drain iterator to be assigned to. Doing so can leave
it with a pointer to a Vec created after it was, and then it has a
dangling pointer in its destructor. Thus was caught by ASAN when I added
a test doing move-assignment.

Add traits on Vec and make forward_decl.h

Vec traits can specify the allocator in the future, for now the specify
the growth factor of the Vec.

Changing the decl of Vec means changing random parts of the codebase,
so to avoid redoing that for other types move all forward decls into
a single place in lib/__private/forward_decl.h

Add drop_iterator_invalidation_tracking to Slice types to allow a Slice
type to simply be detached from invalidation tracking. Drain can do this
since the Slice points into a Vec that it controls and never mutates
the capacity of.

Iterator invalidation has been added to SliceIter, SliceMutIter, Chunks
iters, Split iters, and Windows iters.