Use GCC 13 in the CI bots to avoid GCC 12 bug.
On GCC 12, a mutable variable can't be used in a constexpr context, but
it now works in GCC 13.

Unfortunately, (or fortunately?) GCC 13 reports an OOB write in fmtlib's
float formatting. We disable that warning for now so that we can build,
and it is reported upstream at https://github.com/fmtlib/fmt/issues/3533