Naming for Option and/or/xor with ..._that()
Since these are reserved keywords we can not simply name them as
such unfortunately.

Previously we used the placeholder names and_opt, or_opt, xor_opt, to
indicate it is an operation with another optional. Now we'll move to
naming them in the sense of "this and that", as `this` is how C++
refers to the self pointer.

So we get and_that(Option), or_that(Option), and xor_that(Option).

Closes #132.