Make all of Option constexpr
We just discovered construct_at, which is described as:

Equivalent to
```
return ::new (static_cast<void*>(p)) T(std::forward<Args>(args)...);
```
except that construct_at may be used in evaluation of constant
expressions.

This means we can _change_ the active union field in the Option inside
a constexpr context!