Improve the use of marker types
Ensure operator== and operator<=> have a non-templated overload
against the same type so that markers can convert into the exact
type when possible.

Ensure markers don't convert a reference into a temporary and then
construct a reference out of that. This is done with the new
concept SafelyConstructibleFromReference. To do this for the Choice
marker type we needed to refactor and construct a way to get the
Choice storage types without knowing the actual Choice type.