Add a doxygen config file (it does not work though)
The latest Windows release comes with a Clang that is too old. So you
need to build libclang.dll yourself and drop it into the doxygen
directory.

With the current config at least, it does not chase from header files
into .inc files correctly, so all the documentation for numerics and
slice methods are missing.

For #287