Rename ToBits to Transmogrify
In lack of a better name, we defer to Calvin and Hobbes to show
us how to convert things to other types, through the Transmogrifier.

The sus::mog<T>(f) operation will convert an object `f` to an
object of type T, performing a copy as needed.

The mog operation is supported for subspace numeric types
both integers and floats, primitive integer and float types,
and enums and enum classes.

The operation can be applied to other types by providing a
specialization of the TransmogrifyImpl<To, From> struct with
a To mog_from(const From&) static method.